Working on the orders program.
We can log in.
Roles are implemented.
We change the "names" of the attributes on the model property.
Now we use the ModelSerializer to send the response.
We have the apis to consume warehouses, branches, and orders.
In the orders we have problems with the special characters and, because warehouse and branch are optional, the url not satisfies the request.